[COM-245][JH] Fixed issue of pre-existing component files that were causing issues.
