+++ Core/Core.vcxproj
@@ -156,8 +156,8 @@
     </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="AudioComponent.cpp" />
     <ClCompile Include="src\Component.cpp" />
+    <ClCompile Include="src\Components\AudioComponent.cpp" />
     <ClCompile Include="src\Components\RigidBodyComponent.cpp" />
     <ClCompile Include="src\Components\TransformComponent.cpp" />
     <ClCompile Include="src\Core.cpp" />
@@ -179,8 +179,8 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="AudioComponent.h" />
     <ClInclude Include="include\Component.h" />
+    <ClInclude Include="include\Components\AudioComponent.h" />
     <ClInclude Include="include\Components\RigidBodyComponent.h" />
     <ClInclude Include="include\Components\TransformComponent.h" />
     <ClInclude Include="include\Core.h" />
